refactor:separate test
